\docType{functions}
\name{moment}
\alias{iav}
\alias{moment}
\title{moment}
\usage{
  moment(x, n = rep(1, length(x)), na.rm = T)
}
\arguments{
  \item{x;}{a vector holding a time series}
}
\value{
  a \code{vector} with the inter-annual variation each time
  step
}
\description{
  aa
}
\examples{
x=1
}

